creating the destroyoninteract base class
