+++ Assets/Scripts/Interactable/DestroyOnInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61023c8a1d4d1c04991e47dde3a0a682
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
